edits to get tests working for #93, new test_bellman.py
